+++ scripts/!adetailer.py
@@ -101,8 +101,6 @@ class AfterDetailerScript(scripts.Script):
         self.ultralytics_device = self.get_ultralytics_device()
 
         self.controlnet_ext = None
-        self.cn_script = None
-        self.cn_latest_network = None
 
     def __repr__(self):
         return f"{self.__class__.__name__}(version={__version__})"
@@ -308,6 +306,12 @@ class AfterDetailerScript(scripts.Script):
             sampler_name = "Euler"
         return sampler_name
 
+    def get_override_settings(self, p, args: ADetailerArgs) -> dict[str, Any]:
+        d = {}
+        if args.ad_use_clip_skip:
+            d["CLIP_stop_at_last_layers"] = args.ad_clip_skip
+        return d
+
     def get_initial_noise_multiplier(self, p, args: ADetailerArgs) -> float | None:
         if args.ad_use_noise_multiplier:
             return args.ad_noise_multiplier
@@ -349,9 +353,6 @@ class AfterDetailerScript(scripts.Script):
             filename = Path(filepath).stem
             if filename in script_names_set:
                 filtered_alwayson.append(script_object)
-            if filename == "controlnet":
-                self.cn_script = script_object
-                self.cn_latest_network = script_object.latest_network
 
         script_runner.alwayson_scripts = filtered_alwayson
         return script_runner, script_args
@@ -374,6 +375,7 @@ class AfterDetailerScript(scripts.Script):
         cfg_scale = self.get_cfg_scale(p, args)
         initial_noise_multiplier = self.get_initial_noise_multiplier(p, args)
         sampler_name = self.get_sampler(p, args)
+        override_settings = self.get_override_settings(p, args)
 
         i2i = StableDiffusionProcessingImg2Img(
             init_images=[image],
@@ -409,6 +411,7 @@ class AfterDetailerScript(scripts.Script):
             extra_generation_params=p.extra_generation_params,
             do_not_save_samples=True,
             do_not_save_grid=True,
+            override_settings=override_settings,
         )
 
         i2i.cached_c = [None, None]